% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.htol.R
\name{autoplot.htol}
\alias{autoplot.htol}
\title{Autoplot htol objects}
\usage{
\method{autoplot}{htol}(object, ...)
}
\arguments{
\item{object}{Result from \code{psiiht}}

\item{\dots}{Extra arguments, currently unused}
}
\description{
Default plots for htol objects made by \code{psiiht}
}
