% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.htol.R
\name{fortify.htol}
\alias{fortify.htol}
\title{Fortify htol objects}
\usage{
\method{fortify}{htol}(model, data = model$data, ...)
}
\arguments{
\item{model}{htol object make by \code{psiiht()}}

\item{data}{Orginal data - provided by default}

\item{\dots}{Extra arguments, currently unused}
}
\description{
Process htol objects made by \code{psiiht()} for \code{autoplot()}
}
