% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{htdata}
\alias{htdata}
\title{Heat treatment data}
\format{
\subsection{\code{htdata}}{

A data frame with 720 rows and 6 columns:
\describe{
\item{id}{plant ID}
\item{temperature}{Experimental temperature}
\item{sample.number}{Sample number}
\item{fo}{fo}
\item{fm}{fm}
\item{fvfm}{fvfm}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
htdata
}
\description{
Some heat treatment data
}
\keyword{datasets}
