% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.htol.R
\name{summary.htol}
\alias{summary.htol}
\title{Summary of htol objects}
\usage{
\method{summary}{htol}(object, ...)
}
\arguments{
\item{object}{htol object make by \code{psiiht()}}

\item{\dots}{Extra arguments, currently unused}
}
\description{
Makes summary of htol objects made by \code{psiiht()}
}
