% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSII_heat_tolerance.R
\name{psiiht}
\alias{psiiht}
\title{Heat tolerances}
\usage{
psiiht(temperature, fvfm, id, control_temp, warming = TRUE, boots)
}
\arguments{
\item{temperature}{Temperature}

\item{fvfm}{fvfm data}

\item{id}{leaf ID}

\item{control_temp}{Control temperature}

\item{warming}{Is the experiment analysing heat tolerance (TRUE)
(fvfm declines with high values) or cold tolerance (FALSE)
(fmfv declines with low values)? Important for calculating Tcrit.}

\item{boots}{Number of bootstrap iterations}
}
\description{
Code for estimating heat tolerances using variable fluorescence!
}
\examples{
htol <- psiiht(
  temperature = htdata$temperature, fvfm = htdata$fvfm,
  control_temp = 23, id = htdata$id, boots = 5
)
summary(htol)
autoplot(htol)
}
